fixed from int to Long, and now it works to do quiz and answer the questions. Now i have to fix the score

+++ quizservice/src/main/java/org/quizapp/quizservice/dtos/QuizDTO.java
@@ -6,7 +6,7 @@ import lombok.Data;
 @Data
 @AllArgsConstructor
 public class QuizDTO {
-    private int quizId;
+    private Long quizId;
     private String title;
     private String chapter;
     private String subject;
